Implement transaction support for key-value store
Added a 'Transaction' struct and 'TransactionStack' struct in the 'kvs/transaction.go' allowing for the key-value store to support transactions. This is reflected in 'main.go', where the shell now runs transactions instead of direct operations on the 'KeyValueStore'. Additional error messages were also added in 'kvs/key_value_store.go' to support the transaction error handling. This feature helps in ensuring data consistency and provides greater flexibility in managing data on a session basis.