Improve key retrieval in the transaction process
This commit modifies the Get(key) method in transaction.go to correctly utilize the kvStore object for data retrieval. Previously, the Get() method gave an "ErrKeyNotFound" error when the key was not in the recent changes of a transaction. Now, it will also search in the main store before returning an error. Furthermore, we also remove an outdated comment in the Commit() method.